Move Object to animals.ts
