+++ src/features/game/types/animals.ts
@@ -1,4 +1,6 @@
+import Decimal from "decimal.js-light";
 import { BuildingName } from "./buildings";
+import { AnimalFoodName, Inventory } from "./game";
 
 export type AnimalBuildingType = Extract<BuildingName, "Barn" | "Hen House">;
 
@@ -11,6 +13,14 @@ type AnimalDetail = {
   height: number;
   width: number;
 };
+
+export interface Feed {
+  name: AnimalFoodName;
+  description: string;
+  ingredients: Inventory;
+  coins?: number;
+  disabled?: boolean;
+}
 export const ANIMALS: Record<AnimalType, AnimalDetail> = {
   Chicken: {
     coins: 50,
@@ -52,5 +62,38 @@ export const ANIMAL_LEVELS: Record<AnimalType, Record<AnimalLevel, number>> = {
     1: 0,
     2: 20,
     3: 50,
+  },
+};
+
+export const ANIMAL_FOODS: Record<AnimalFoodName, Feed> = {
+  Hay: {
+    name: "Hay",
+    description: "",
+    ingredients: {
+      Corn: new Decimal(1),
+    },
+  },
+  "Kernel Blend": {
+    name: "Kernel Blend",
+    description: "",
+    ingredients: {
+      Wheat: new Decimal(1),
+    },
+  },
+  NutriBarley: {
+    name: "NutriBarley",
+    description: "",
+    ingredients: {
+      Barley: new Decimal(1),
+    },
+  },
+  "Mixed Grain": {
+    name: "Mixed Grain",
+    description: "",
+    ingredients: {
+      Wheat: new Decimal(1),
+      Corn: new Decimal(1),
+      Barley: new Decimal(1),
+    },
   },
 };
